requirements.txt
semgrep
ansi2html
python-owasp-zap-v2.4